frontend init

+++ backend/controllers/product.controller.js
@@ -23,7 +23,9 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
-    console.log("id:",id);
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message: "Invalid product id"});
+    }
     try{
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product deleted successfully"});
